refactor(post): 상세 페이지 라우트 파라미터를 postId로 변경
