+++ app/[locale]/blog/[nickname]/[[...slug]]/page.tsx
@@ -1,1 +1,1 @@
-export { default } from "@/app/post/[id]/PostDetailPage";
+export { default } from "@/app/post/[postId]/PostDetailPage";
