--- app/[locale]/post/[id]/page.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/app/post/[id]/PostDetailPage";
