+++ app/[locale]/post/[postId]/page.tsx
